+++ src/module/data/pseudo-documents/advancements/_module.mjs
@@ -1,4 +1,5 @@
 export { default as BaseAdvancement } from "./base-advancement.mjs";
+export { default as CharacteristicAdvancement } from "./characteristic.mjs";
 export { default as ItemGrantAdvancement } from "./item-grant-advancement.mjs";
 export { default as LanguageAdvancement } from "./language-advancement.mjs";
 export { default as SkillAdvancement } from "./skill-advancement.mjs";